Move `croud login` tests to dedicated module
