+++ tests/test_authentication.py
@@ -17,169 +17,11 @@
 # with Crate these terms will supersede the license and you may use the
 # software solely pursuant to the terms of the relevant commercial agreement.
 
-import copy
 from argparse import Namespace
-from typing import Dict
 from unittest import mock
 
-import pytest
-
-from croud.config import Configuration
-from croud.login import _login_url, get_org_id, login
 from croud.logout import _logout_url, logout
-from croud.server import Server
-
-
-class MockConfig:
-    """
-    A mocked configuration which emulates reading and writing file from/to disk.
-    """
-
-    def __init__(self, config: Dict) -> None:
-        self.conf = copy.deepcopy(config)
-
-    def read_config(self) -> Dict:
-        return self.conf
-
-    def write_config(self, config) -> None:
-        self.conf = config
-
-
-class TestLogin:
-    @mock.patch.object(Server, "stop")
-    @mock.patch.object(Server, "start")
-    @mock.patch("croud.login.asyncio.get_event_loop")
-    @mock.patch("croud.login.can_launch_browser", return_value=True)
-    @mock.patch("croud.login.open_page_in_browser")
-    @mock.patch("croud.login.print_info")
-    def test_login_success(
-        self,
-        mock_print_info,
-        mock_open_page_in_browser,
-        mock_can_launch_browser,
-        mock_loop,
-        mock_start,
-        mock_stop,
-    ):
-        cfg = MockConfig(Configuration.DEFAULT_CONFIG)
-
-        with mock.patch("croud.config.load_config", side_effect=cfg.read_config):
-            with mock.patch("croud.config.write_config", side_effect=cfg.write_config):
-                with mock.patch("croud.login.get_org_id", return_value="my-org-id"):
-                    login(Namespace(env="dev"))
-
-        calls = [
-            mock.call("A browser tab has been launched for you to login."),
-            mock.call("Login successful."),
-        ]
-        mock_print_info.assert_has_calls(calls)
-        config = cfg.read_config()
-        assert config["auth"]["current_context"] == "dev"
-        assert config["auth"]["contexts"]["dev"]["organization_id"] == "my-org-id"
-
-    @mock.patch.object(Server, "stop")
-    @mock.patch.object(Server, "start")
-    @mock.patch("croud.login.asyncio.get_event_loop")
-    @mock.patch("croud.login.can_launch_browser", return_value=True)
-    @mock.patch("croud.login.open_page_in_browser")
-    @mock.patch("croud.login.print_info")
-    def test_local_login(
-        self,
-        mock_print_info,
-        mock_open_page_in_browser,
-        mock_can_launch_browser,
-        mock_loop,
-        mock_start,
-        mock_stop,
-    ):
-        """
-        Test for a bug that caused that upon login to local env the local token
-        was overwritten by the dev token.
-        """
-        conf = {
-            "auth": {
-                "current_context": "prod",
-                "contexts": {
-                    "prod": {"token": "prod-token"},
-                    "dev": {"token": "dev-token"},
-                    "local": {"token": ""},
-                },
-            },
-            "region": "bregenz.a1",
-            "output_fmt": "table",
-        }
-        cfg = MockConfig(conf)
-
-        with mock.patch("croud.config.load_config", side_effect=cfg.read_config):
-            with mock.patch("croud.config.write_config", side_effect=cfg.write_config):
-                with mock.patch("croud.rest.Client.send", return_value=[None, None]):
-                    login(Namespace(env="local"))
-
-        new_cfg = cfg.read_config()
-        assert new_cfg["auth"]["current_context"] == "local"
-        assert (
-            new_cfg["auth"]["contexts"]["local"]["token"]
-            != new_cfg["auth"]["contexts"]["dev"]["token"]
-        )
-        assert (
-            new_cfg["auth"]["contexts"]["local"]["token"]
-            != new_cfg["auth"]["contexts"]["prod"]["token"]
-        )
-
-    @mock.patch("croud.config.Configuration.get_env", return_value="dev")
-    @mock.patch("croud.login.can_launch_browser", return_value=False)
-    @mock.patch("croud.login.print_error")
-    def test_login_no_valid_browser(
-        self, mock_print_error, mock_can_launch_browser, mock_get_env
-    ):
-        with pytest.raises(SystemExit) as e_info:
-            login(Namespace(env=None))
-
-        mock_print_error.assert_called_once_with(
-            "Login only works with a valid browser installed."
-        )
-        assert e_info.value.code == 1
-
-    def test_login_urls_from_valid_envs(self):
-        url = _login_url("dev")
-        assert "https://bregenz.a1.cratedb-dev.cloud/oauth2/login?cli=true" == url
-
-        url = _login_url("prod")
-        assert "https://bregenz.a1.cratedb.cloud/oauth2/login?cli=true" == url
-
-        url = _login_url("PROD")
-        assert "https://bregenz.a1.cratedb.cloud/oauth2/login?cli=true" == url
-
-        url = _login_url("local")
-        assert "http://localhost:8000/oauth2/login?cli=true" == url
-
-    def test_env_fallback_url(self):
-        url = _login_url("invalid")
-        assert "https://bregenz.a1.cratedb.cloud/oauth2/login?cli=true" == url
-
-    def test_get_org_id_no_org(self):
-        cfg = MockConfig(Configuration.DEFAULT_CONFIG)
-        cfg.conf["auth"]["contexts"]["dev"]["organization_id"]
-        with mock.patch("croud.config.load_config", side_effect=cfg.read_config):
-            with mock.patch("croud.config.write_config", side_effect=cfg.write_config):
-                with mock.patch(
-                    "croud.rest.Client.send",
-                    return_value=[{"organization_id": None}, None],
-                ):
-                    org_id = get_org_id()
-        assert org_id is None
-
-    def test_get_org_id_org(self):
-        cfg = MockConfig(Configuration.DEFAULT_CONFIG)
-        cfg.conf["auth"]["contexts"]["dev"]["organization_id"]
-        with mock.patch("croud.config.load_config", side_effect=cfg.read_config):
-            with mock.patch("croud.config.write_config", side_effect=cfg.write_config):
-                with mock.patch(
-                    "croud.rest.Client.send",
-                    return_value=[{"organization_id": "some_id"}, None],
-                ):
-                    org_id = get_org_id()
-        assert org_id == "some_id"
+from tests.util import MockConfig
 
 
 class TestLogout:
